+++ src/components/forms/NetworkDevicesForm.tsx
@@ -12,8 +12,8 @@ import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "util/queryKeys";
 import { fetchNetworks } from "api/networks";
 import { LxdNicDevice } from "types/device";
-import { SharedFormikTypes } from "pages/instances/forms/sharedFormTypes";
-import InstanceConfigurationTable from "pages/instances/forms/InstanceConfigurationTable";
+import { SharedFormikTypes } from "./sharedFormTypes";
+import InstanceConfigurationTable from "components/forms/InstanceConfigurationTable";
 import { fetchProfiles } from "api/profiles";
 import { EditInstanceFormValues } from "pages/instances/EditInstanceForm";
 import { getConfigurationRowBase } from "components/ConfigurationRow";
