+++ src/components/forms/SecurityPoliciesForm.tsx
@@ -7,12 +7,9 @@ import {
   optionTrueFalse,
   optionYesNo,
 } from "util/instanceOptions";
-import {
-  SharedFormikTypes,
-  SharedFormTypes,
-} from "pages/instances/forms/sharedFormTypes";
-import { getInstanceConfigurationRow } from "pages/instances/forms/InstanceConfigurationRow";
-import InstanceConfigurationTable from "pages/instances/forms/InstanceConfigurationTable";
+import { SharedFormikTypes, SharedFormTypes } from "./sharedFormTypes";
+import { getInstanceConfigurationRow } from "components/forms/InstanceConfigurationRow";
+import InstanceConfigurationTable from "components/forms/InstanceConfigurationTable";
 import { getInstanceKey } from "util/instanceConfigFields";
 import { optionRenderer } from "util/formFields";
