+++ src/pages/profiles/EditProfileForm.tsx
@@ -19,20 +19,20 @@ import { FormDeviceValues, formDeviceToPayload } from "util/formDevices";
 import SecurityPoliciesForm, {
   SecurityPoliciesFormValues,
   securityPoliciesPayload,
-} from "pages/instances/forms/SecurityPoliciesForm";
+} from "components/forms/SecurityPoliciesForm";
 import SnapshotsForm, {
   SnapshotFormValues,
   snapshotsPayload,
-} from "pages/instances/forms/SnapshotsForm";
+} from "components/forms/SnapshotsForm";
 import CloudInitForm, {
   CloudInitFormValues,
   cloudInitPayload,
-} from "pages/instances/forms/CloudInitForm";
+} from "components/forms/CloudInitForm";
 import ResourceLimitsForm, {
   ResourceLimitsFormValues,
   resourceLimitsPayload,
-} from "pages/instances/forms/ResourceLimitsForm";
-import YamlForm, { YamlFormValues } from "pages/instances/forms/YamlForm";
+} from "components/forms/ResourceLimitsForm";
+import YamlForm, { YamlFormValues } from "components/forms/YamlForm";
 import { updateProfile } from "api/profiles";
 import ProfileFormMenu, {
   CLOUD_INIT,
@@ -47,8 +47,8 @@ import ProfileFormMenu, {
 import { LxdProfile } from "types/profile";
 import useEventListener from "@use-it/event-listener";
 import { updateMaxHeight } from "util/updateMaxHeight";
-import DiskDeviceForm from "pages/instances/forms/DiskDeviceForm";
-import NetworkDevicesForm from "pages/instances/forms/NetworkDevicesForm";
+import DiskDeviceForm from "components/forms/DiskDeviceForm";
+import NetworkDevicesForm from "components/forms/NetworkDevicesForm";
 import ProfileDetailsForm, {
   profileDetailPayload,
   ProfileDetailsFormValues,
