+++ src/util/instanceEdit.tsx
@@ -4,10 +4,10 @@ import { formDeviceToPayload, parseDevices } from "util/formDevices";
 import { parseCpuLimit, parseMemoryLimit } from "util/limits";
 import { getInstanceConfigKeys } from "util/instanceConfigFields";
 import { instanceEditDetailPayload } from "pages/instances/forms/InstanceEditDetailsForm";
-import { resourceLimitsPayload } from "pages/instances/forms/ResourceLimitsForm";
-import { securityPoliciesPayload } from "pages/instances/forms/SecurityPoliciesForm";
-import { snapshotsPayload } from "pages/instances/forms/SnapshotsForm";
-import { cloudInitPayload } from "pages/instances/forms/CloudInitForm";
+import { resourceLimitsPayload } from "components/forms/ResourceLimitsForm";
+import { securityPoliciesPayload } from "components/forms/SecurityPoliciesForm";
+import { snapshotsPayload } from "components/forms/SnapshotsForm";
+import { cloudInitPayload } from "components/forms/CloudInitForm";
 import { getUnhandledKeyValues } from "util/formFields";
 import { EditInstanceFormValues } from "pages/instances/EditInstanceForm";
 import * as Yup from "yup";
